refactor: domain 모듈에서 token을 user 하위로 넣고 설정 패키지 변경
