+++ devridge-core/domain/src/main/java/io/devridge/core/domain/user/TokenRepository.java
@@ -1,4 +1,4 @@
-package io.devridge.core.domain.token;
+package io.devridge.core.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
